fix: event detection
The different webhook events that might be configured at admin or
per-repo (user) level are not consistent in the use of the
'object_kind' key to identify them, thus detection fails.

Improve detection by testing against other keys that may identify the
event, and default to a non identified message (this will be used to
report that condition).