improve: repository_update handling
formatRepoUpdateMsg was trying to do too much and badly:

tried to ignore the case where a single tag push or remove was the
only operation in place, since it's supposed to be handled only by
formatTagPushMsg, but only if there weren't any other operations in
place

we also need to ignore head (branch) changes that are caused by normal
commit pushes, which do update branches

This is doable in an orderly fashion with a double pass, but in the end
it's better to actually have tag operations reported at two places,
because repository update webhooks can only be enabled by site admins
and there are cases when this setting won't be changed by them. Besides,
the reporting format of the two formatters differ (one is atomic, the
other aggregates all changes in one notification).