feature: chat state
Add methods and structures to maintain and refresh information about the
bot activity.

self.owners = list of bot owners
self.chats = list of chats where the bot is active
self.otp = list of active, issued OTP tokens
self.challenges = list of OTP challenges currently active

The first three lists are also stored as part of self.state, which is
itself a member of self.config, thus serialized by save_config()

These are lists of dictionaries that always have a 'refresh' entry,
storing a timestamp in the future after which the entry is to be
refreshed. The process of refreshing is particular to each list, and
for particular entries might mean updating dictionary members, or
removing the entry altogether.

refresh() is the implementation of the class virtual method, going
through the different lists

update_chat() is used to initialize (when called by cache_chat()) and to
update chat info (when called by refresh())

cache_chat()
a simple method that returns the chat information from cache if there is
any, or initializes / caches the information before returning it
otherwise

user_entry()
this generates the dictionary storing information about a user